Clarify atan2 overloads
In the C++ standard, this function is only defined for floating point
types, and integers cannot be implicitly converted. Using explicit
conversions avoids a GCC specific extension to the standard library.

+++ gerbview/gerbview_frame.cpp
@@ -755,7 +755,7 @@ void GERBVIEW_FRAME::UpdateStatusBar()
         dy = GetCrossHairPosition().y - screen->m_O_Curseur.y;
 
         // atan2 in the 0,0 case returns 0
-        theta = RAD2DEG( atan2( -dy, dx ) );
+        theta = RAD2DEG( atan2( (double) -dy, (double) dx ) );
 
         ro = hypot( dx, dy );
         wxString formatter;
